+++ src/viewsets/ProjectInviteViewSet.ts
@@ -14,8 +14,6 @@ import { requirePermission } from '../authorization/middleware';
 import { transaction } from 'objection';
 import validate from '../utils/validate';
 
-const sendMail = config.MAIL_BACKEND.sendMail;
-
 const projectInviteHTML = mjml2html(
   fs.readFileSync(
     path.resolve(__dirname, '../../templates/mail/project-invite.mjml'), 'utf8'
@@ -122,7 +120,7 @@ export default class ProjectInviteViewSet extends BaseViewSet<ProjectInvite> {
       text: ejs.render(projectInviteTXT, context),
       html: ejs.render(projectInviteHTML, context)
     };
-    sendMail(mailOptions)
+    this.options.mail.sendMail(mailOptions)
       .then(info => {
         console.log('Message sent: %s', info.messageId); // tslint:disable-line
       })